make better use of given models and abstractions
