+++ cert_issuer/blockchain_handlers/ethereum_sc/transaction_handlers.py
@@ -2,7 +2,7 @@ import logging
 
 from pycoin.serialize import b2h
 
-from cert_issuer.blockchain_handlers.ethereum_sc.connections import ContractConnection
+from cert_issuer.blockchain_handlers.ethereum_sc.connectors import EthereumSCServiceProviderConnector
 from cert_issuer.errors import InsufficientFundsError
 from cert_issuer.blockchain_handlers.ethereum_sc import tx_utils
 from cert_issuer.models import TransactionHandler
@@ -39,9 +39,9 @@ class EthereumSCTransactionHandler(TransactionHandler):
     def issue_transaction(self, blockchain_bytes, app_config):
         eth_data_field = b2h(blockchain_bytes)
 
-        self.contract = ContractConnection(app_config)
+        # self.contract = ContractConnection(app_config)
 
-        prepared_tx = self.contract.functions.create_transaction("issue_hash", blockchain_bytes)
+        prepared_tx = self.connector.create_transaction("issue_hash", blockchain_bytes)
         signed_tx = self.sign_transaction(prepared_tx)
 
         # TODO
@@ -49,20 +49,6 @@ class EthereumSCTransactionHandler(TransactionHandler):
 
         txid = self.broadcast_transaction(signed_tx)
         return txid
-
-    def create_transaction(self, blockchain_bytes):
-        if self.balance:
-            # it is assumed here that the address has sufficient funds, as the ensure_balance has just been checked
-            nonce = self.connector.get_address_nonce(self.issuing_address)
-            # Transactions in the first iteration will be send to burn address
-            toaddress = '0xdeaddeaddeaddeaddeaddeaddeaddeaddeaddead'
-            tx = self.transaction_creator.create_transaction(self.tx_cost_constants, self.issuing_address, nonce,
-                                                             toaddress, blockchain_bytes)
-
-            prepared_tx = tx
-            return prepared_tx
-        else:
-            raise InsufficientFundsError('Not sufficient ether to spend at: %s', self.issuing_address)
 
     def sign_transaction(self, prepared_tx):
         # stubbed from BitcoinTransactionHandler
@@ -74,7 +60,7 @@ class EthereumSCTransactionHandler(TransactionHandler):
         return signed_tx
 
     def broadcast_transaction(self, signed_tx):
-        txid = self.contract.functions.broadcast_tx(signed_tx)
+        txid = self.connector.broadcast_tx(signed_tx)
         return txid
 
     def verify_transaction(self, signed_tx, eth_data_field):
