+++ test_setup.sh
@@ -3,4 +3,4 @@ mkdir -p venv/
 rm -r venv/test/
 virtualenv venv/test
 source venv/test/bin/activate
-python setup.py install experimental --blockchain=ethereum
+python setup.py install experimental --blockchain=ethereum_smart_contract
